% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laber_test.R
\name{est_pows}
\alias{est_pows}
\title{A helper function to estimate the power using the generated test statistic,
and estimated distribution of the test statistic.}
\usage{
est_pows(tr_lm_dstr, ts_vec)
}
\arguments{
\item{tr_lm_dstr}{limiting distribution of the test statistic}

\item{ts_vec}{Vector containing the test statistic}
}
\value{
Return the estimated power of the test based for
each of the possible sum of squares norms
}
\description{
A helper function to estimate the power using the generated test statistic,
and estimated distribution of the test statistic.
}
