% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laber_test.R
\name{get_test_stat}
\alias{get_test_stat}
\title{Helper function for the Zhang and Laber test.}
\usage{
get_test_stat(obs_data)
}
\arguments{
\item{obs_data}{The observed data}
}
\value{
Returns the estimated t-statistics for the Zhang and Laber tests.
}
\description{
Helper function for the Zhang and Laber test.
}
