% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mag_for_pow.R
\name{mag_for_pow}
\alias{mag_for_pow}
\title{This function is used to estimate the magnitude needed for a
certain direction to achieve 80\% power for a proposed alternative.}
\usage{
mag_for_pow(mc_limit_dstr, dir, norms_idx = 2, null_quants, norm_type = "lp")
}
\arguments{
\item{mc_limit_dstr}{MC draws from the estimated (centered)
limiting distribution. Columns correspond to different covariates,
and rows are independent observations.}

\item{dir}{the direction for which we wish to find the magnitude needed
to achieve 80\% power.}

\item{norms_idx}{the norms to be considered.}

\item{null_quants}{ninety five percent quantiles for each of the
different norms}

\item{norm_type}{the class of norms to be used}
}
\value{
Magnitude for the specified lp norm for a given local alternative.
}
\description{
This function is used to estimate the magnitude needed for a
certain direction to achieve 80\% power for a proposed alternative.
}
