% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_for_mag.R
\name{pval_for_mag}
\alias{pval_for_mag}
\title{A helper function that calculates the estimated p-value for a given
observed alternative and a given norm.}
\usage{
pval_for_mag(mc_limit_dstr, dir, norms_idx = 2, norm_type = "lp", ...)
}
\arguments{
\item{mc_limit_dstr}{the simulated data draw from
the limiting distribution under the null}

\item{dir}{the observed estimate of the parameter}

\item{norms_idx}{the index for the norm used}

\item{norm_type}{the type of norm used}

\item{...}{additional arguments that may be passed to
pval_for_mag, but which will be ignored.}
}
\value{
The p-value for a test with estimate given by
\code{dir} and limiting distribution given by \code{mc_limit_dstr}.
}
\description{
A helper function that calculates the estimated p-value for a given
observed alternative and a given norm.
}
